FITRUN-113 refactor: api, domain 패키지 분리
