+++ src/main/kotlin/com/yapp/yapp/audio/domain/AudioRepository.kt
@@ -1,4 +1,4 @@
-package com.yapp.yapp.audio
+package com.yapp.yapp.audio.domain
 
 interface AudioRepository {
     fun getRandomCoachAudio(): AudioResource
